+++ pulsar/net/__init__.py
@@ -4,6 +4,7 @@ called ``stream``. It can be assessed by::
 
     from pulsar import net
 '''
+from .sock import *
 from .base import *
 from .tcp import *
 from .http import *