+++ pulsar/utils/sock/__init__.py
@@ -1,3 +1,0 @@
-from .base import *
-from .plat import *
-from .create import *