+++ pulsar/utils/sock/plat.py
@@ -1,6 +1,0 @@
-from pulsar.utils.system import platform
-
-if platform.type == 'posix':
-    from .posix import *
-elif platform.type == 'win':
-    from .win import *